fix: Added support for runtiveConfig from envs
