+++ playground/nuxt.config.ts
@@ -3,6 +3,13 @@ import MyModule from '../'
 
 export default defineNuxtConfig({
   modules: [MyModule],
+  runtimeConfig: {
+    public: {
+      yandexMetrika: {
+        id: '', // NUXT_PUBLIC_YANDEX_METRIKA_ID
+      },
+    },
+  },
   yandexMetrika: {
     id: '49439650',
   },
